incorporating CR's text feedback:
(from email)

Under Context, the second half of the sentence below reads awkwardly. Mind fixing? Perhaps using systems language like, 'the department detains them disproportionately'? The sentence after it with, "Youth continue to be stopped..." might also benefit from flipping to something like, 'The department continues to stop youth...'

"Fresno has a large Southeast Asian population that feels the effects of how the department creates harm by gang profiling, as well as Latinx men who experience higher number of detainments as a result of traffic stops."

Also under Context, can we get any kind of ballpark figure on the sentence below? It would make a more impactful point in context. This is also the first recommendation and in the conclusions. Even a sense of $1 million, $10million, etc would go a long way to make the recommendations more actionable.

"Furthermore, Fresno PD’s commitment to such unproductive and harmful practices annually wastes millions of public dollars."

Under Methodology or the top of Findings, it would be helpful to state the total number of stops and people stopped in the data analyzed. This would set up when later findings are sometimes presented by stops other times by people stopped. Larger question, many of the findings charts by race but some charts don't have total figures presented with them for comparison (others have totals in pop ups). Is there a reason for that? Totals don't need to be on the charts themselves but could be helpful in surrounding paragraphs or even in an appendix.

Finding 1. Half of people are (to match the second half of the sentence)?

"Approximately half (50.5%) of people living in Fresno city is Latinx, while a quarter (25.4%) are White"